Create no-op Window pathlocks implementation
This keeps the call sites simple, eventually this should be filled in.
