+++ src/build-remote/build-remote.cc
@@ -22,6 +22,7 @@
 #include "experimental-features.hh"
 
 using namespace nix;
+using namespace nix::unix;
 using std::cin;
 
 static void handleAlarm(int sig) {
