+++ src/libstore/lock.cc
@@ -9,6 +9,8 @@
 
 namespace nix {
 
+using namespace nix::unix;
+
 #if __linux__
 
 static std::vector<gid_t> get_group_list(const char *username, gid_t group_id)
