+++ src/libstore/windows/pathlocks-impl.hh
@@ -1,0 +1,2 @@
+#pragma once
+///@file Needed because Unix-specific counterpart
